misc: share component for taxes selection
