+++ src/components/plans/CommitmentsSection.tsx
@@ -5,19 +5,16 @@ import { Icon } from 'lago-design-system'
 import { RefObject, useEffect, useMemo, useState } from 'react'
 
 import { Accordion, Button, Chip, Tooltip, Typography } from '~/components/designSystem'
-import { AmountInputField, ComboBox, ComboboxItem } from '~/components/form'
+import { AmountInputField } from '~/components/form'
 import { EditInvoiceDisplayNameDialogRef } from '~/components/invoices/EditInvoiceDisplayNameDialog'
 import { PremiumWarningDialogRef } from '~/components/PremiumWarningDialog'
-import {
-  MUI_INPUT_BASE_ROOT_CLASSNAME,
-  SEARCH_TAX_INPUT_FOR_MIN_COMMITMENT_CLASSNAME,
-} from '~/core/constants/form'
+import { TaxesSelectorSection } from '~/components/taxes/TaxesSelectorSection'
+import { SEARCH_TAX_INPUT_FOR_MIN_COMMITMENT_CLASSNAME } from '~/core/constants/form'
 import { intlFormatNumber } from '~/core/formats/intlFormatNumber'
-import { scrollToAndClickElement } from '~/core/utils/domUtils'
 import {
   CommitmentTypeEnum,
   CurrencyEnum,
-  useGetTaxesForCommitmentsLazyQuery,
+  TaxForTaxesSelectorSectionFragmentDoc,
 } from '~/generated/graphql'
 import { useInternationalization } from '~/hooks/core/useInternationalization'
 import { useCurrentUser } from '~/hooks/useCurrentUser'
@@ -34,10 +31,12 @@ gql`
       }
       collection {
         id
-        ...TaxForPlanUsageChargeAccordion
+        ...TaxForTaxesSelectorSection
       }
     }
   }
+
+  ${TaxForTaxesSelectorSectionFragmentDoc}
 `
 
 type CommitmentsSectionProps = {
@@ -54,48 +53,11 @@ export const CommitmentsSection = ({
   const { isPremium } = useCurrentUser()
   const { translate } = useInternationalization()
 
-  const [shouldDisplayTaxesInput, setShouldDisplayTaxesInput] = useState<boolean>(false)
   const [displayMinimumCommitment, setDisplayMinimumCommitment] = useState<boolean>(
     !isNaN(Number(formikProps.initialValues.minimumCommitment?.amountCents)),
   )
 
   const hasErrorInGroup = !!formikProps?.errors?.minimumCommitment
-
-  const [getTaxes, { data: taxesData, loading: taxesLoading }] = useGetTaxesForCommitmentsLazyQuery(
-    {
-      variables: { limit: 500 },
-    },
-  )
-  const { collection: taxesCollection } = taxesData?.taxes || {}
-
-  const taxesDataForCombobox = useMemo(() => {
-    if (!taxesCollection) return []
-
-    const minCommitmentsTaxesIds =
-      formikProps.values.minimumCommitment?.taxes?.map((tax) => tax.id) || []
-
-    return taxesCollection.map(({ id: taxId, name, rate }) => {
-      const formatedRate = intlFormatNumber(Number(rate) / 100 || 0, {
-        style: 'percent',
-      })
-
-      return {
-        label: `${name} (${formatedRate})`,
-        labelNode: (
-          <ComboboxItem>
-            <Typography variant="body" color="grey700" noWrap>
-              {name}
-            </Typography>
-            <Typography variant="caption" color="grey600" noWrap>
-              {formatedRate}
-            </Typography>
-          </ComboboxItem>
-        ),
-        value: taxId,
-        disabled: minCommitmentsTaxesIds.includes(taxId),
-      }
-    })
-  }, [formikProps.values.minimumCommitment?.taxes, taxesCollection])
 
   const taxValueForBadgeDisplay = useMemo((): string | undefined => {
     if (!formikProps?.values?.minimumCommitment?.taxes?.length) return
@@ -212,89 +174,17 @@ export const CommitmentsSection = ({
               <Typography className="mb-2" variant="captionHl" color="grey700">
                 {translate('text_64be910fba8ef9208686a8e3')}
               </Typography>
-              <div className="flex flex-col gap-4">
-                {!!formikProps?.values?.minimumCommitment?.taxes?.length && (
-                  <div className="flex flex-wrap items-center gap-3">
-                    {formikProps?.values?.minimumCommitment?.taxes.map(
-                      ({ id: localTaxId, name, rate }) => (
-                        <Chip
-                          key={localTaxId}
-                          label={`${name} (${rate}%)`}
-                          type="secondary"
-                          size="medium"
-                          deleteIcon="trash"
-                          icon="percentage"
-                          deleteIconLabel={translate('text_63aa085d28b8510cd46443ff')}
-                          onDelete={() => {
-                            const newTaxedArray =
-                              formikProps?.values?.minimumCommitment?.taxes?.filter(
-                                (tax) => tax.id !== localTaxId,
-                              ) || []
 
-                            formikProps.setFieldValue('minimumCommitment.taxes', newTaxedArray)
-                          }}
-                        />
-                      ),
-                    )}
-                  </div>
-                )}
-
-                {shouldDisplayTaxesInput ? (
-                  <div className="flex items-center gap-3">
-                    <ComboBox
-                      containerClassName="flex-1"
-                      className={SEARCH_TAX_INPUT_FOR_MIN_COMMITMENT_CLASSNAME}
-                      data={taxesDataForCombobox}
-                      searchQuery={getTaxes}
-                      loading={taxesLoading}
-                      placeholder={translate('text_64be910fba8ef9208686a8e7')}
-                      emptyText={translate('text_64be91fd0678965126e5657b')}
-                      onChange={(newTaxId) => {
-                        const previousTaxes = [
-                          ...(formikProps?.values?.minimumCommitment?.taxes || []),
-                        ]
-                        const newTaxObject = taxesData?.taxes?.collection.find(
-                          (t) => t.id === newTaxId,
-                        )
-
-                        formikProps.setFieldValue('minimumCommitment.taxes', [
-                          ...previousTaxes,
-                          newTaxObject,
-                        ])
-                        setShouldDisplayTaxesInput(false)
-                      }}
-                    />
-
-                    <Tooltip placement="top-end" title={translate('text_63aa085d28b8510cd46443ff')}>
-                      <Button
-                        icon="trash"
-                        variant="quaternary"
-                        onClick={() => {
-                          setShouldDisplayTaxesInput(false)
-                        }}
-                      />
-                    </Tooltip>
-                  </div>
-                ) : (
-                  // Wrapping div to avoid the button to be full width, caused by the <Stack> parent
-                  <div>
-                    <Button
-                      fitContent
-                      startIcon="plus"
-                      variant="inline"
-                      onClick={() => {
-                        setShouldDisplayTaxesInput(true)
-
-                        scrollToAndClickElement({
-                          selector: `.${SEARCH_TAX_INPUT_FOR_MIN_COMMITMENT_CLASSNAME} .${MUI_INPUT_BASE_ROOT_CLASSNAME}`,
-                        })
-                      }}
-                    >
-                      {translate('text_64be910fba8ef9208686a8c9')}
-                    </Button>
-                  </div>
-                )}
-              </div>
+              <TaxesSelectorSection
+                taxes={formikProps?.values?.minimumCommitment?.taxes || []}
+                comboboxSelector={SEARCH_TAX_INPUT_FOR_MIN_COMMITMENT_CLASSNAME}
+                onUpdate={(newTaxArray) => {
+                  formikProps.setFieldValue('minimumCommitment.taxes', newTaxArray)
+                }}
+                onDelete={(newTaxArray) => {
+                  formikProps.setFieldValue('minimumCommitment.taxes', newTaxArray)
+                }}
+              />
             </div>
           </Stack>
         </Accordion>
