+++ src/components/plans/PlanSettingsSection.tsx
@@ -1,27 +1,18 @@
 import { gql } from '@apollo/client'
 import { FormikProps } from 'formik'
-import { memo, useEffect, useMemo, useState } from 'react'
+import { memo, useEffect, useState } from 'react'
 
-import { Button, Card, Chip, Tooltip, Typography } from '~/components/designSystem'
-import {
-  ButtonSelectorField,
-  ComboBox,
-  ComboBoxField,
-  ComboboxItem,
-  TextInput,
-  TextInputField,
-} from '~/components/form'
+import { Button, Card, Tooltip, Typography } from '~/components/designSystem'
+import { ButtonSelectorField, ComboBoxField, TextInput, TextInputField } from '~/components/form'
+import { TaxesSelectorSection } from '~/components/taxes/TaxesSelectorSection'
 import {
   FORM_ERRORS_ENUM,
   FORM_TYPE_ENUM,
   getIntervalTranslationKey,
-  MUI_INPUT_BASE_ROOT_CLASSNAME,
   SEARCH_TAX_INPUT_FOR_PLAN_CLASSNAME,
 } from '~/core/constants/form'
-import { intlFormatNumber } from '~/core/formats/intlFormatNumber'
-import { scrollToAndClickElement } from '~/core/utils/domUtils'
 import { updateNameAndMaybeCode } from '~/core/utils/updateNameAndMaybeCode'
-import { CurrencyEnum, PlanInterval, useGetTaxesForPlanLazyQuery } from '~/generated/graphql'
+import { CurrencyEnum, PlanInterval } from '~/generated/graphql'
 import { useInternationalization } from '~/hooks/core/useInternationalization'
 
 import { PlanFormInput } from './types'
@@ -82,40 +73,7 @@ export const PlanSettingsSection = memo(
     const [shouldDisplayDescription, setShouldDisplayDescription] = useState<boolean>(
       !!formikProps.initialValues.description,
     )
-    const [shouldDisplayTaxesInput, setShouldDisplayTaxesInput] = useState(false)
     const plan = formikProps.values
-    const [getTaxes, { data: taxesData, loading: taxesLoading }] = useGetTaxesForPlanLazyQuery({
-      variables: { limit: 20 },
-    })
-    const { collection: taxesCollection } = taxesData?.taxes || {}
-
-    const taxesDataForCombobox = useMemo(() => {
-      if (!taxesCollection) return []
-
-      const planTaxesIds = plan.taxes?.map((tax) => tax.id) || []
-
-      return taxesCollection.map(({ id, name, rate }) => {
-        const formatedRate = intlFormatNumber(Number(rate) / 100 || 0, {
-          style: 'percent',
-        })
-
-        return {
-          label: `${name} (${formatedRate})`,
-          labelNode: (
-            <ComboboxItem>
-              <Typography variant="body" color="grey700" noWrap>
-                {name}
-              </Typography>
-              <Typography variant="caption" color="grey600" noWrap>
-                {formatedRate}
-              </Typography>
-            </ComboboxItem>
-          ),
-          value: id,
-          disabled: planTaxesIds.includes(id),
-        }
-      })
-    }, [plan.taxes, taxesCollection])
 
     useEffect(() => {
       setShouldDisplayDescription(!!formikProps.initialValues.description)
@@ -238,88 +196,24 @@ export const PlanSettingsSection = memo(
           name="amountCurrency"
         />
 
-        {!!plan?.taxes?.length && (
-          <div>
-            <Typography className="mb-1" variant="captionHl" color="grey700">
+        <div className="flex flex-col gap-1">
+          {!!plan?.taxes?.length && (
+            <Typography variant="captionHl" color="grey700">
               {translate('text_6661fc17337de3591e29e3e1')}
             </Typography>
-            <div className="flex flex-wrap items-center gap-3">
-              {plan.taxes.map(({ id, name, rate }) => (
-                <Chip
-                  key={id}
-                  label={`${name} (${rate}%)`}
-                  size="medium"
-                  deleteIcon="trash"
-                  icon="percentage"
-                  deleteIconLabel={
-                    subscriptionFormType === FORM_TYPE_ENUM.edition || (isEdition && !canBeEdited)
-                      ? undefined
-                      : translate('text_63aa085d28b8510cd46443ff')
-                  }
-                  onDelete={() => {
-                    const newTaxedArray = plan.taxes?.filter((tax) => tax.id !== id) || []
+          )}
 
-                    formikProps.setFieldValue('taxes', newTaxedArray)
-                  }}
-                />
-              ))}
-            </div>
-          </div>
-        )}
-
-        {shouldDisplayTaxesInput ? (
-          <div>
-            {!plan.taxes?.length && (
-              <Typography className="mb-1" variant="captionHl" color="grey700">
-                {translate('text_6661fc17337de3591e29e3e1')}
-              </Typography>
-            )}
-            <div className="flex items-center gap-3">
-              <ComboBox
-                containerClassName="flex-1"
-                className={SEARCH_TAX_INPUT_FOR_PLAN_CLASSNAME}
-                data={taxesDataForCombobox}
-                searchQuery={getTaxes}
-                loading={taxesLoading}
-                placeholder={translate('text_6661fc17337de3591e29e3e3')}
-                emptyText={translate('text_64be91fd0678965126e5657b')}
-                onChange={(newTaxId) => {
-                  const previousTaxes = [...(formikProps?.values?.taxes || [])]
-                  const newTaxObject = taxesData?.taxes?.collection.find((t) => t.id === newTaxId)
-
-                  formikProps.setFieldValue('taxes', [...previousTaxes, newTaxObject])
-                  setShouldDisplayTaxesInput(false)
-                }}
-              />
-
-              <Tooltip placement="top-end" title={translate('text_63aa085d28b8510cd46443ff')}>
-                <Button
-                  icon="trash"
-                  variant="quaternary"
-                  onClick={() => {
-                    setShouldDisplayTaxesInput(false)
-                  }}
-                />
-              </Tooltip>
-            </div>
-          </div>
-        ) : (
-          <Button
-            className="self-start"
-            startIcon="plus"
-            variant="inline"
-            onClick={() => {
-              setShouldDisplayTaxesInput(true)
-
-              scrollToAndClickElement({
-                selector: `.${SEARCH_TAX_INPUT_FOR_PLAN_CLASSNAME} .${MUI_INPUT_BASE_ROOT_CLASSNAME}`,
-              })
+          <TaxesSelectorSection
+            taxes={plan.taxes || []}
+            comboboxSelector={SEARCH_TAX_INPUT_FOR_PLAN_CLASSNAME}
+            onUpdate={(newTaxArray) => {
+              formikProps.setFieldValue('taxes', newTaxArray)
             }}
-            data-test="show-add-taxes"
-          >
-            {translate('text_64be910fba8ef9208686a8c9')}
-          </Button>
-        )}
+            onDelete={(newTaxArray) => {
+              formikProps.setFieldValue('taxes', newTaxArray)
+            }}
+          />
+        </div>
       </Card>
     )
   },
