+++ translations/base.json
@@ -951,7 +951,6 @@
   "text_6661fc17337de3591e29e3d1": "Billing period",
   "text_6661fc17337de3591e29e3d3": "Frequency at which the subscription recurs.",
   "text_6661fc17337de3591e29e3e1": "Tax rates",
-  "text_6661fc17337de3591e29e3e3": "Select a tax rate",
   "text_6661fc17337de3591e29e3e7": "Pricing settings",
   "text_6661fc17337de3591e29e3e9": "Customize the pricing structure of the plan.",
   "text_6661fc17337de3591e29e3ed": "{{interval}} recurring amount paid by the customer regardless of their usage consumption.",
