+++ translations/pt-BR.json
@@ -951,7 +951,6 @@
   "text_6661fc17337de3591e29e3d1": "Período de faturamento",
   "text_6661fc17337de3591e29e3d3": "Frequência com que a assinatura se repete.",
   "text_6661fc17337de3591e29e3e1": "Alíquotas de imposto",
-  "text_6661fc17337de3591e29e3e3": "Selecione uma alíquota de imposto",
   "text_6661fc17337de3591e29e3e7": "Configurações de preço",
   "text_6661fc17337de3591e29e3e9": "Personalize a estrutura de preços do plano.",
   "text_6661fc17337de3591e29e3ed": "Valor recorrente {{interval}} pago pelo cliente, independentemente do consumo de uso.",
